refactor: ビルド設定の関連クラスを apps/app/lib/configuration へ移動
