+++ apps/app/lib/app_initializer.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
-import 'package:cores_core/provider.dart';
 import 'package:cores_core/util.dart';
 import 'package:flutter_app/app_build_config.dart';
 import 'package:flutter_app/composition_root/data_sources/shared_preference_data_source.dart';
+import 'package:flutter_app/presentation/providers/build_config_provider.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
