+++ apps/app/lib/configuration/flavor.dart
