--- packages/cores/core/lib/model.dart
@@ -1,2 +1,0 @@
-export 'src/model/build_config.dart';
-export 'src/model/flavor.dart';
