--- packages/cores/core/lib/provider.dart
@@ -1,1 +1,0 @@
-export 'src/provider/build_config_provider.dart';
